+++ General/Misc.vb
@@ -3844,12 +3844,6 @@ Public Enum CompressionMode
     none
 End Enum
 
-Public Enum x265Type
-    Vanilla
-    DJATOM
-    Asuna
-End Enum
-
 Public Enum ApplicationEvent
     <DispName("After Job Muxed")> JobMuxed
     <DispName("After Job Processed")> JobProcessed